# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MTB REST API
  version: 1.0.0
  description: REST API of the MTB service.
...
servers:
  - url: /api
paths:
...
  /episodes/{episodeId}/diagnoses:

    post:
      summary: add diagnose for an episode.
      description: add diagnose for an episode.
      operationId: addDiagnose
      tags:
        - diagnose
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
      requestBody:
        description: add diagnose for an episode.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Diagnose'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Diagnose'
        400:
          description: Bad Request - Foreign key constraint violated (EpisodeId or one of histologyResultId not found) or invalid germline diagnose
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: Episode not found

    get:
      summary: List all available diagnoses for an episode.
      operationId: getAllDiagnoses
      tags:
        - diagnose
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Diagnose'
        400:
          description: Foreign key constraint violated (Episode is not the same as the saved in the database)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
...
  /episodes/{episodeId}/specimens:

    post:
      summary: add specimen for an episode.
      description: add specimen for an episode.
      operationId: addSpecimen
      tags:
        - specimen
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
      requestBody:
        description: add specimen for an episode.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Specimen'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specimen'
        400:
          description: Foreign key constraint violated (EpisodeId not found) or body and path EpisodeId variables do not match
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        # 409:
        #   description: Diagnose for the given Episode with the given date already exists

    get:
      summary: List all available specimens for an episode.
      operationId: getAllSpecimens
      tags:
        - specimen
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Specimen'
        400:
          description: Foreign key constraint violated (Episode is not the same as the saved in the database)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
  #        404:
  #          description: Specimen not found
  /episodes/{episodeId}/specimens/{specimenId}:

    get:
      summary: get specimen information for a specific specimenId.
      operationId: getSpecimen
      tags:
        - specimen
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
        - name: specimenId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the Specimen
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specimen'
        400:
          description: Foreign key constraint violated (Episode not found or is not the same as the saved in the database)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: Specimen not found
    #        409:
    #          description: Specimen for the given Episode with the given date already exists

    put:
      summary: Change specimen for an episode.
      description: Change specimen for an episode.
      operationId: updateSpecimen
      tags:
        - specimen
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
        - name: specimenId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the Specimen
      requestBody:
        description: Change Specimen for an episode.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Specimen'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specimen'
        400:
          description: Foreign key constraint violated (Provided Episode does not match the one associated with the Specimen)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: Specimen not found
        405:
          description: Invalid Input

    delete:
      summary: Delete specimen for an episode.
      description: Delete specimen for an episode.
      operationId: deleteSpecimen
      tags:
        - specimen
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
        - name: specimenId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the Specimen
      responses:
        200:
          description: OK
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: Specimen not found
...
components:
  schemas:
...
    Diagnose:
      type: object
      properties:
        id:
          type: string
          format: uuid
        episodeId:
          type: string
          format: uuid
        recordedOn:
          type: string
          format: date
          example: '2020-05-27'
        icd10:
          $ref: '#/components/schemas/Code'
        icdO3T:
          $ref: '#/components/schemas/Code'
        whoGrade:
          $ref: '#/components/schemas/Code'
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/StatusHistory'
        guidelineTreatmentStatus:
          type: string
          enum: [ "exhausted", "non-exhausted", "impossible",  "no-guidelines-available", "unknown" ]
        histologyResults:
          type: array
          items:
            type: string
            format: uuid
        isGermlineDiagnosisExist:
          type: boolean
        germlineDiagnosisIcd10:
          $ref: '#/components/schemas/Code'
        hpoIcd10:
          type: array
          items:
            $ref: '#/components/schemas/Code'
        orphanetCode:
          $ref: '#/components/schemas/Code'
        alphaIdSeCode:
          $ref: '#/components/schemas/Code'
        tnmKey:
          $ref: '#/components/schemas/Code'
        altTumorKey:
          type: array
          items:
            $ref: '#/components/schemas/Code'
    Code:
      type: object
      properties:
        code:
          type: string
        display:
          type: string
        version:
          type: string
        system:
          type: string
          # enum: [ICD10, ICDO3T, WHO_GRADE]
...
    StatusHistory:
      type: object
      properties:
        status:
          type: string
          enum: [ "tumor-free", "local", "metastasized", "unknown" ]
        date:
          type: string
          format: date
    Specimen:
      type: object
      properties:
        id:
          type: string
          format: uuid
        labelling:
          type: string
        episodeId:
          type: string
          format: uuid
        icd10:
          $ref: '#/components/schemas/Code'
        type:
          type: string
          enum: [ "fresh-tissue", "cryo-frozen", "liquid-biopsy", "FFPE", "unknown" ]
        collection:
          $ref: '#/components/schemas/SpecimenCollection'
    SpecimenCollection:
      type: object
      properties:
        localization:
          type: string
          enum: [ "primary-tumor", "metastasis" , "unknown", "local-recurrence", "regional-lymph-node", "cellfree-dna" ]
        date:
          type: string
          format: date
        method:
          type: string
          enum: [ "biopsy", "resection", "liquid-biopsy", "cytology", "unknown" ]
...